AD-HOC feat (Smart Cards): Add role to install smart card daemon
I tend to use a Smart Card (specifically a Yubikey) for a whole swathe
of cryptographic operations, including signing commits, SSH, various
other bits and bobs.

This commit installs a daemon required for GPG to read information from
this smart card.

== Design Notes ==

=== Yubikey Permissions Issues ===

Currently there is still an outstanging bug in which the unprivileged
GnuPG user can't read the Yubikey device. This makes sense, as the
hardware shouldn't be readable by all users, but is a bit annoying as it
means that I can't use it at the minute.

Need some thinking.