AD-HOC refactor (Smart Card): Replace smart card role with yubikey role
Previously there was installed some packages to allow reading from Smart
Cards. The only smart card in known use is the Yubikey 4; requiring
additional configuration in order to function -- namely modification of
udev rules to allow systemd to determine access to the device.

These rules are packaged by Debian in the package "libu2f-host0".
This package is Yubikey specific, but still solves the aforementioned
problems with permissions.

The generic role for smart cards was made a more specific role for
yubikeys, and the additional package appended to this role.

See https://wiki.debian.org/Smartcards/YubiKey4

== Design Notes ==

=== Why not separate roles? ===

While it would have been defensible to create a separate role for the
Yubikey, the roles are so simple it seems wasteful to create another.