AD-HOC feat (PHPStorm): Include the PHPStorm role
PHPStorm is the canonical development environment for the author. This
commit uses the existing and seemingly well crafted pixelart.phpstorm
role to manage this installation.

== Design Notes ==

=== Snapd? ===

Snapd is a new system to manage the installation of PHPStorm. However,
there does not appear to be good ansible primitives exposed that
allow the management of these packages.