AD-HOC feat (X): Add the xclip tool
The primary window manager for Debian stretch+ is the Xorg window
manager. It provides a tool called "xclip" which takes content off the
command line and puts it in the clipboard managed by X. This is
extremely useful when going back and fourth between graphics and CLI
applications as a mechanism to store data in a temporary place.

This commit install this application.

== Design Notes ==

=== Separate Role ===

This commit establishes the xorg role. Future work may customize things
like mouse pointer behaviour from this role.